+++ packages/types/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ansospace/types
 
+## 0.3.8
+
+### Patch Changes
+
+- Renamed `mongooseObjectId` to `objectId` and `MongooseObjectId` to `ObjectId` in `common.ts`.
+- Updated `objectId` schema to use `ObjectId` from "bson" for validation and transformation.
+
 ## 0.3.7
 
 ### Minor Changes
@@ -34,7 +41,7 @@
 
 ### Patch Changes
 
-- Updated `mongooseObjectId` schema in `common.ts` to use `mongoose.Types.ObjectId.isValid` for validation and transform strings into `mongoose.Types.ObjectId` instances.
+- Updated `objectId` schema in `common.ts` to use `mongoose.Types.ObjectId.isValid` for validation and transform strings into `mongoose.Types.ObjectId` instances.
 - Fixed type incompatibility issues between client-side string `id` and Mongoose `ObjectId` by aligning validation and transformation logic.
 
 ## 0.3.1